Add feature to detect if PDFs were created digitally
New flag, -d, will direct the program to check the RGB values for
seven pixels which are found in the thin, black borders between cells
in various positions around the rendered page of the PDF. Scanned
documents are exceptionally unlikely to have each of these pixels have
a perfectly black pixel in each location, given they are generally
created by scanning a paper document. If all pixels are found to be
entirely black, then it is a reasonable assumption to believe the PDF
was generated from keyed values and that the crop extents in the program
will reliably extract the narrative and diagram.