Add argument to have the crash diagram saved to local disk
The argument --save-diagram is renamed to --save-diagram-s3.
An argument --save-diagram-disk is added that needs the path, in the
form of a directory, to which the crash diagram is saved.

This is done to support the accumulation of crash diagrams for
inspection to assess the degree to which PII may be included.

Example invocation: ./cr3_extract_diagram_ocr_narrative.py -v -d \
--cr3-source atd-vision-zero-editor production/cris-cr3-files \
--save-diagram-disk /tmp/crash_diagrams --crash-id <crash id>